requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
authlib==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
supabase==2.15.3
schedule==1.2.1
python-dateutil==2.8.2
aiohttp==3.11.18
jinja2==3.1.3
starlette==0.27.0
pydantic==1.10.13
redis==5.0.1
dramatiq==1.15.0
